Stubbed a basic framework for the NimbusApplication.
One weakness of this arrangement is that it aggregates a lot of
functionality in a few functions. However, this makes tracing the line of
execution along the base framework more simple, so I think it's a good
idea.

As we develop more of the framework, we should consider decentralizing a
lot of the initialization stuff that this base class does. (e.g. scene
creation should be delegated to run mode and OIS input handling should be
handled by the input manager.)