Create singleton parent class.
This is a working copy, but it exhibits problems I believe are inherent in
the SelfInitializingSingleton class.

Unless someone else can fix them, I'd
recommend that we abondon this idea as a general rule and only implement
one or two singleton classes.

I think that NimbusApplication could be our singleton on which everything
else could be built. If we need other singleton type objects, we could add
them to NimbusApplication as members and access them through accessor
methods... That seems more acceptable to me, but it may still lead to
similar problems if not carefully managed.