Add InputManager with OIS configured.
Functions basically as the InputManager should. This version lacks all
mechanisms for binding and messaging. (Particularly because no messaging
system exists.)

In order to prove that input is working, debug messages are printed to the
console with mouse button and key presses. Additionally, the escape key
has been enabled to kill the application (though this is done through a
hack... in the future, this should be accomplished through the
messaging/event system).