Document our mouse click events
Our mouse click events currently do no
delegation of where a click is meant to
be handled. This does not need to send
an empty payload, so for now, it does not.
A payload will be re-added when we have
proper delegation of what defines who
should handle a click.