Add listeners, wind plane collision, and arrows
Listeners are implemented to create points and
append the drag points of the mouse to our
current working current (cwc) (as opposed to
current working directory, cwd) which is then
passed off to WindMap.
Wind plane collision is a utility function to
return the 2-dimensional point at which a click
collides with our wind plane. Yay for code
cleanliness.
Arrows were deprecated and then readded in a
stroke of purr awesome. Typo intentional.